API 인증 관련 수정 (#25)
* fix, refactor: SecurityConfig 수정

[refactor]
- @EnableMethodSecurity : PrePostAuthorize으로 메소드 단위 접근 권한 제어
- @EnableWebSecurity : @EnableMethod 의존, 추후 OAuth2 확장 고려
- @securedEnabled : 메소드 @Secured 설정 가능
- authorizeRequest : 명시적 허용 선언으로 수정

[fix]
- 순환 참조 방지 PasswordEncoder SecurityConfig → AppConfig 이동

* fix: JwtAuthFilter 수정

* chore: 프론트 디폴트 포트 수정
