+++ src/main/java/com/ll/eitcharge/global/security/SecurityConfig.java
@@ -2,45 +2,41 @@ package com.ll.eitcharge.global.security;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.security.config.annotation.method.configuration.EnableMethodSecurity;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
-import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
-import org.springframework.security.crypto.password.PasswordEncoder;
+import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
+import org.springframework.security.config.annotation.web.configurers.AbstractHttpConfigurer;
+import org.springframework.security.config.http.SessionCreationPolicy;
 import org.springframework.security.web.SecurityFilterChain;
+import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
+
+import lombok.RequiredArgsConstructor;
 
 @Configuration
+@RequiredArgsConstructor
+@EnableWebSecurity
+@EnableMethodSecurity(securedEnabled = true)
 public class SecurityConfig {
-    @Bean
-    public PasswordEncoder passwordEncoder() {
-        return new BCryptPasswordEncoder();
-    }
+    private final JwtAuthenticationFilter jwtAuthenticationFilter;
 
     @Bean
-    public SecurityFilterChain filterChain(HttpSecurity http) throws Exception {
+    SecurityFilterChain filterChain(HttpSecurity http) throws Exception {
         http
-                .authorizeRequests(authorizeRequests ->
-                        authorizeRequests
-                                .requestMatchers("/gen/**")
-                                .permitAll()
-                                .requestMatchers("/resource/**")
-                                .permitAll()
-                                .requestMatchers("/h2-console/**")
-                                .permitAll()
-                                .anyRequest()
-                                .permitAll()
-                )
-                .headers(
-                        headers ->
-                                headers.frameOptions(
-                                        frameOptions ->
-                                                frameOptions.sameOrigin()
-                                )
-                )
-                .csrf(
-                        csrf ->
-                                csrf.ignoringRequestMatchers(
-                                        "/h2-console/**"
-                                )
-                );
+            .authorizeRequests(authorizeRequests ->
+                authorizeRequests
+                    .requestMatchers("/gen/**")
+                    .permitAll()
+            )
+            .csrf(
+                AbstractHttpConfigurer::disable
+            )
+            .sessionManagement(
+                sessionManagement -> sessionManagement
+                    .sessionCreationPolicy(
+                        SessionCreationPolicy.STATELESS
+                    )
+            )
+            .addFilterBefore(jwtAuthenticationFilter, UsernamePasswordAuthenticationFilter.class);
 
         return http.build();
     }
